refactor: collapse XbehaveDelegateInvoker down into StepInvoker
