+++ src/Xbehave.2.Execution.Net45/Xbehave.2.Execution.Net45.csproj
@@ -62,7 +62,6 @@
     <Compile Include="Shims\ISerializable.cs" />
     <Compile Include="Shims\SerializableAttribute.cs" />
     <Compile Include="StepInvoker.cs" />
-    <Compile Include="XbehaveDelegateInvoker.cs" />
     <Compile Include="StepRunner.cs" />
     <Compile Include="TypeInfoExtensions.cs" />
     <Compile Include="ScenarioOutlineRunner.cs" />
