--- src/Xbehave.2.Execution.Net45/XbehaveDelegateInvoker.cs
@@ -1,84 +1,0 @@
-﻿// <copyright file="XbehaveDelegateInvoker.cs" company="xBehave.net contributors">
-//  Copyright (c) xBehave.net contributors. All rights reserved.
-// </copyright>
-
-namespace Xbehave.Execution
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
-    using System.Reflection;
-    using System.Security;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Xbehave.Execution.Shims;
-    using Xunit.Abstractions;
-    using Xunit.Sdk;
-
-    public abstract class XbehaveDelegateInvoker : XunitTestInvoker
-    {
-        protected XbehaveDelegateInvoker(
-            ITest test,
-            IMessageBus messageBus,
-            Type testClass,
-            object[] constructorArguments,
-            MethodInfo testMethod,
-            object[] testMethodArguments,
-            IReadOnlyList<BeforeAfterTestAttribute> beforeAfterAttributes,
-            ExceptionAggregator aggregator,
-            CancellationTokenSource cancellationTokenSource)
-            : base(
-                test,
-                messageBus,
-                testClass,
-                constructorArguments,
-                testMethod,
-                testMethodArguments,
-                beforeAfterAttributes,
-                aggregator,
-                cancellationTokenSource)
-        {
-        }
-
-        public async override Task<decimal> InvokeTestMethodAsync(object testClassInstance)
-        {
-            var oldSyncContext = SynchronizationContext.Current;
-
-            try
-            {
-                var asyncSyncContext = new AsyncTestSyncContext(oldSyncContext);
-                SetSynchronizationContext(asyncSyncContext);
-
-                await Aggregator.RunAsync(
-                    () => Timer.AggregateAsync(
-                        async () =>
-                        {
-                            await this.InvokeDelegatesAsync();
-                            var ex = await asyncSyncContext.WaitForCompletionAsync();
-                            if (ex != null)
-                            {
-                                Aggregator.Add(ex);
-                            }
-                        }));
-            }
-            finally
-            {
-                SetSynchronizationContext(oldSyncContext);
-            }
-
-            return Timer.Total;
-        }
-
-        protected abstract Task InvokeDelegatesAsync();
-
-        [SuppressMessage(
-            "Microsoft.Security",
-            "CA2136:TransparencyAnnotationsShouldNotConflictFxCopRule",
-            Justification = "From xunit.")]
-        [SecuritySafeCritical]
-        private static void SetSynchronizationContext(SynchronizationContext context)
-        {
-            SynchronizationContext.SetSynchronizationContext(context);
-        }
-    }
-}
